refactor and reduce
